pyproject.toml
[tool.poetry]
name = "oidx_mds"
version = "0.1.0"
description = ""
authors = ["Sean Davis <seandavi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
httpx = "^0.23.0"
anyio = "^3.6.2"
uvicorn = "^0.19.0"
pendulum = "^2.1.2"
omicidx = "1.5.1"
google-cloud-secret-manager = "^2.12.6"

[tool.poetry.dev-dependencies]
mypy = "^0.982"
black = "^22.10.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
